Modify entrypoint

+++ ppml/trusted-big-data-ml/python/docker-gramine/base/entrypoint.sh
@@ -110,18 +110,35 @@ case "$SPARK_K8S_CMD" in
     if [ "$SGX_ENABLED" == "false" ]; then
         $SPARK_HOME/bin/spark-submit --conf spark.driver.bindAddress=$SPARK_DRIVER_BIND_ADDRESS --deploy-mode client "$@"
     elif [ "$SGX_ENABLED" == "true" ]; then
-	export driverExtraClassPath=`cat /opt/spark/conf/spark.properties | grep -P -o "(?<=spark.driver.extraClassPath=).*"` && \
+	      export driverExtraClassPath=`cat /opt/spark/conf/spark.properties | grep -P -o "(?<=spark.driver.extraClassPath=).*"` && \
         echo $driverExtraClassPath && \
         export SGX_MEM_SIZE=$SGX_DRIVER_MEM_SIZE && \
         export sgx_command="/opt/jdk8/bin/java -Dlog4j.configurationFile=/ppml/trusted-big-data-ml/work/spark-3.1.2/conf/log4j2.xml -Xms1G -Xmx$SGX_DRIVER_JVM_MEM_SIZE -cp "$SPARK_CLASSPATH:$driverExtraClassPath" org.apache.spark.deploy.SparkSubmit --conf spark.driver.bindAddress=$SPARK_DRIVER_BIND_ADDRESS --deploy-mode client "$@"" && \
         if [ "$ATTESTATION" = "true" ]; then
-	  bash attestation.sh
-	  echo $sgx_command >> temp_command_file
-	  export sgx_command="bash temp_command_file && rm temp_command_file"
-	fi
+          # Also consider ENCRYPTEDFSD condition
+          rm /ppml/temp_command_file || true
+	        bash attestation.sh
+          if [ "$ENCRYPTED_FSD" == "true" ]; then
+            echo "[INFO] Distributed encrypted file system is enabled"
+            bash encrypted-fsd.sh
+          fi
+          echo $sgx_command >> temp_command_file
+          export sgx_command="bash temp_command_file && rm temp_command_file"
+	      fi
+        # ATTESTATION is false
+        if [ "$ENCRYPTED_FSD" == "true" ]; then
+          # ATTESTATION false, encrypted-fsd true
+          rm /ppml/temp_command_file || true
+          echo "[INFO] Distributed encrypted file system is enabled"
+          bash encrypted-fsd.sh
+          echo $sgx_command >> temp_command_file
+          export sgx_command="bash temp_command_file && rm temp_command_file"
+        fi
+        # Attestation false, encrypted-fsd false
         echo $sgx_command && \
         ./init.sh && \
-	gramine-sgx bash  1>&2
+	      gramine-sgx bash  2>&1
+        rm /ppml/temp_command_file || true
     fi
     ;;
   driver-py)
@@ -167,13 +184,29 @@ case "$SPARK_K8S_CMD" in
       export SGX_MEM_SIZE=$SGX_EXECUTOR_MEM_SIZE && \
       export sgx_command="/opt/jdk8/bin/java -Dlog4j.configurationFile=/ppml/trusted-big-data-ml/work/spark-3.1.2/conf/log4j2.xml -Xms1G -Xmx$SGX_EXECUTOR_JVM_MEM_SIZE "${SPARK_EXECUTOR_JAVA_OPTS[@]}" -cp "$SPARK_CLASSPATH" org.apache.spark.executor.CoarseGrainedExecutorBackend --driver-url $SPARK_DRIVER_URL --executor-id $SPARK_EXECUTOR_ID --cores $SPARK_EXECUTOR_CORES --app-id $SPARK_APPLICATION_ID --hostname $SPARK_EXECUTOR_POD_IP --resourceProfileId $SPARK_RESOURCE_PROFILE_ID" && \
       if [ "$ATTESTATION" = "true" ]; then
-        bash attestation.sh
-	echo $sgx_command >> temp_command_file
-	export sgx_command="bash temp_command_file && rm temp_command_file"
-      fi
-      echo $sgx_command && \
+          # Also consider ENCRYPTEDFSD condition
+          rm /ppml/temp_command_file || true
+	        bash attestation.sh
+          if [ "$ENCRYPTED_FSD" == "true" ]; then
+            echo "[INFO] Distributed encrypted file system is enabled"
+            bash encrypted-fsd.sh
+          fi
+          echo $sgx_command >> temp_command_file
+          export sgx_command="bash temp_command_file && rm temp_command_file"
+	      fi
+        # ATTESTATION is false
+        if [ "$ENCRYPTED_FSD" == "true" ]; then
+          # ATTESTATION false, encrypted-fsd true
+          rm /ppml/temp_command_file || true
+          echo "[INFO] Distributed encrypted file system is enabled"
+          bash encrypted-fsd.sh
+          echo $sgx_command >> temp_command_file
+          export sgx_command="bash temp_command_file && rm temp_command_file"
+        fi
+        # Attestation false, encrypted-fsd false
+        echo $sgx_command && \
       ./init.sh && \
-      gramine-sgx bash  1>&2
+      gramine-sgx bash  2>&1
     fi
     ;;
